pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

# ─── Project metadata ─────────────────────────────────────────────────────────
[project]
name = "pdf-rag-mcp"
version = "0.1.0"
description = "A fully-local PDF RAG system exposed as an MCP server with a 7-layer hallucination guard stack."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"

dependencies = [
    "pymupdf>=1.24.0",
    "chromadb>=0.5.0",
    "httpx>=0.27.0",
    "mcp[cli]>=1.0.0",
    "streamlit>=1.35.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
# Convenience CLI entrypoints after `pip install -e .`
pdf-rag-server = "src.mcp_server.server:mcp.run"

# ─── Setuptools package discovery ────────────────────────────────────────────
[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

# ─── Ruff (linter + formatter) ────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",  # line-too-long — handled by formatter
    "B008",  # function-calls-in-default-args (FastMCP uses this pattern)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

# ─── Pytest ───────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "-v --tb=short"
asyncio_mode = "auto"
